+++ src/interfaces/template.ts
@@ -1,0 +1,3 @@
+export interface ITemplate {
+  content: string;
+}
