+++ src/resources/number.ts
@@ -1,8 +1,3 @@
-
-export function numberPolyfill() {
-  Number.isInteger = Number.isInteger || ((value) => {
-    return typeof value === "number" &&
-      isFinite(value) &&
-      Math.floor(value) === value;
-  });
-}
+Number.isInteger = Number.isInteger || ((value) => {
+  return typeof value === "number" && isFinite(value) && Math.floor(value) === value;
+});
