+++ tslint.json
@@ -11,6 +11,7 @@
       true,
       "ignore-comments"
     ],
-    "no-string-literal": false
+    "no-string-literal": false,
+    "forin": false
   }
 }
